+++ netpyne/__init__.py
@@ -1,3 +1,3 @@
 
-__version__ = '0.7.0' 
+__version__ = '0.7.1' 
 __gui__ = True  # global option to enable/disable graphics