+++ netpyne/neuromlFuncs.py
@@ -45,7 +45,7 @@ def _convertNetworkRepresentation (net, gids_vs_pop_indices):
                             delay = conn['delay']
                             sec = conn['sec']
                             synMech = conn['synMech']
-                            threshold = conn['threshold']
+                            #threshold = conn['threshold']
 
                             if sim.cfg.verbose: print("      Conn %s[%i]->%s[%i] with %s, w: %s, d: %s"%(popPre, indexPre,popPost, indexPost, synMech, weight, delay))
 
@@ -86,7 +86,7 @@ def _convertStimulationRepresentation (net,gids_vs_pop_indices, nml_doc):
                                 assert(not netstim_found)
                                 netstim_found = True
                                 synMech = conn['synMech']
-                                threshold = conn['threshold']
+                                #threshold = conn['threshold']
                                 delay = conn['delay']
                                 weight = conn['weight']
                                 
@@ -98,11 +98,20 @@ def _convertStimulationRepresentation (net,gids_vs_pop_indices, nml_doc):
                             stims[stim_info] = []
 
 
-                        stims[stim_info].append({'index':index,'weight':weight,'delay':delay,'threshold':threshold})   
+                        stims[stim_info].append({'index':index,'weight':weight,'delay':delay}) 
+                        #stims[stim_info].append({'index':index,'weight':weight,'delay':delay,'threshold':threshold})
 
     #print(stims)
     return stims
 
+#
+#  Heaviside function, required for expressions in <inhomogeneousValue>
+#
+def H(x):
+    if x == 0:
+        return 0.5
+
+    return 1 * (x > 0)
 
 if neuromlExists:
 
@@ -672,12 +681,14 @@ if neuromlExists:
                 prox = parent_seg.distal
 
             dist = seg.distal
-            if prox.x==dist.x and prox.y==dist.y and prox.z==dist.z:
+            
+            # Spherical root segment
+            if seg.parent==None and prox.x==dist.x and prox.y==dist.y and prox.z==dist.z:
 
                 if prox.diameter==dist.diameter:
                     dist.y = prox.diameter
                 else:
-                    raise Exception('Unsupported geometry in segment: %s of cell %s'%(seg.name,cell.id))
+                    raise Exception('Unsupported geometry in segment: %s of cell'%(seg.name))
                 
             return prox, dist
 
@@ -707,7 +718,8 @@ if neuromlExists:
             popInfo['cellsList'] = []
             
             if population_id=='pop':
-                print("\n\n*****************************\nReconsider calling your population 'pop'; it leads to some errors!\n*****************************\n\n")
+                print("\n\n*****************************\nReconsider calling your population 'pop'; it leads to some errors in NetPyNE!\nGiving up...\n*****************************\n\n")
+                quit()
 
             self.popParams[population_id] = popInfo
 
@@ -728,6 +740,16 @@ if neuromlExists:
                                 
                 use_segment_groups_for_neuron = False
                 
+                
+                if len(cell.biophysical_properties.membrane_properties.spike_threshes)>0:
+                    st = cell.biophysical_properties.membrane_properties.spike_threshes[0]
+                    # Ensure threshold is same everywhere on cell
+                    assert(st.segment_groups=='all')
+                    assert(len(cell.biophysical_properties.membrane_properties.spike_threshes)==1)
+                    threshold = pynml.convert_to_units(st.value,'mV')
+                else:
+                    threshold = 0
+                
                 for seg_grp in cell.morphology.segment_groups:
                     if hasattr(seg_grp,'neuro_lex_id') and seg_grp.neuro_lex_id == "sao864921383":
                         use_segment_groups_for_neuron = True
@@ -737,6 +759,8 @@ if neuromlExists:
                                 cellRule['secs'][seg_grp.id]['geom']['nseg'] = int(prop.value)
                         #seg_grps_vs_nrn_sections[seg_grp.id] = seg_grp.id
                         seg_grps_vs_nrn_sections['all'].append(seg_grp.id)
+                        
+                        cellRule['secs'][seg_grp.id]['threshold'] = threshold
                         
                 self.pop_ids_vs_use_segment_groups_for_neuron[population_id] = use_segment_groups_for_neuron
                     
@@ -784,7 +808,7 @@ if neuromlExists:
                                 
                             cellRule['secs'][section]['geom']['pt3d'].append((dist.x,dist.y,dist.z,dist.diameter))
                         
-
+                inhomogeneous_parameters = {}
                     
                 for seg_grp in cell.morphology.segment_groups:
                     seg_grps_vs_nrn_sections[seg_grp.id] = []
@@ -805,9 +829,36 @@ if neuromlExists:
 
                     if not seg_grp.neuro_lex_id or seg_grp.neuro_lex_id !="sao864921383":
                         cellRule['secLists'][seg_grp.id] = seg_grps_vs_nrn_sections[seg_grp.id]
+                        
+                    for ip in seg_grp.inhomogeneous_parameters:
+                        #print("=====================\ninhomogeneousParameter: %s"%ip)
+                        
+                        inhomogeneous_parameters[seg_grp.id] = {}
+                        
+                        ## Some checks here to ensure the defaults/recommended values are selected
+                        # Can be made more general
+                        assert ip.metric=="Path Length from root"
+                        assert ip.variable=="p"
+                        if ip.proximal:
+                            assert float(ip.proximal.translation_start)==0.0
+                        
+                        ordered_segs, cumulative_lengths, path_prox, path_dist = cell.get_ordered_segments_in_groups([seg_grp.id],include_cumulative_lengths=True,include_path_lengths=True)
+                        
+                        nrn_secs = seg_grps_vs_nrn_sections[seg_grp.id]
+                        for nrn_sec in nrn_secs:
+                            sec_segs = cell.get_ordered_segments_in_groups(nrn_sec)
+                            first = sec_segs[nrn_sec][0]
+                            last = sec_segs[nrn_sec][-1]
+                            start_len = path_prox[seg_grp.id][first.id]
+                            end_len = path_dist[seg_grp.id][last.id]
+                            #print("  Seg: %s (%s) -> %s (%s)"%(first,start_len,last,end_len))
+                            
+                            inhomogeneous_parameters[seg_grp.id][nrn_sec] = (start_len,end_len)
+                            
                     
                 
                 for cm in cell.biophysical_properties.membrane_properties.channel_densities:
+                              
                     group = 'all' if not cm.segment_groups else cm.segment_groups
                     for section_name in seg_grps_vs_nrn_sections[group]:
                         gmax = pynml.convert_to_units(cm.cond_density,'S_per_cm2')
@@ -819,12 +870,13 @@ if neuromlExists:
                         
                         cellRule['secs'][section_name]['mechs'][cm.ion_channel] = mech
                         
-                        if cm.ion and cm.ion == 'non_specific':
+                        ion = self._determine_ion(cm)
+                        if ion == 'non_specific':
                             mech['e'] = erev
                         else:
-                            if not cellRule['secs'][section_name]['ions'].has_key(cm.ion):
-                                cellRule['secs'][section_name]['ions'][cm.ion] = {}
-                            cellRule['secs'][section_name]['ions'][cm.ion]['e'] = erev
+                            if not cellRule['secs'][section_name]['ions'].has_key(ion):
+                                cellRule['secs'][section_name]['ions'][ion] = {}
+                            cellRule['secs'][section_name]['ions'][ion]['e'] = erev
                             
                 for cm in cell.biophysical_properties.membrane_properties.channel_density_nernsts:
                     group = 'all' if not cm.segment_groups else cm.segment_groups
@@ -839,22 +891,89 @@ if neuromlExists:
                         
                         #TODO: erev!!
                         
-                        if cm.ion and cm.ion == 'non_specific':
+                        ion = self._determine_ion(cm)
+                        if ion == 'non_specific':
                             pass
                             ##mech['e'] = erev
                         else:
-                            if not cellRule['secs'][section_name]['ions'].has_key(cm.ion):
-                                cellRule['secs'][section_name]['ions'][cm.ion] = {}
-                            ##cellRule['secs'][section_name]['ions'][cm.ion]['e'] = erev
+                            if not cellRule['secs'][section_name]['ions'].has_key(ion):
+                                cellRule['secs'][section_name]['ions'][ion] = {}
+                            ##cellRule['secs'][section_name]['ions'][ion]['e'] = erev
+                            
+                            
+                for cm in cell.biophysical_properties.membrane_properties.channel_density_ghk2s:
+                              
+                    group = 'all' if not cm.segment_groups else cm.segment_groups
+                    for section_name in seg_grps_vs_nrn_sections[group]:
+                        gmax = pynml.convert_to_units(cm.cond_density,'S_per_cm2')
+                        if cm.ion_channel=='pas':
+                            mech = {'g':gmax}
+                        else:
+                            mech = {'gmax':gmax}
+                        
+                        ##erev = pynml.convert_to_units(cm.erev,'mV')
+                        
+                        cellRule['secs'][section_name]['mechs'][cm.ion_channel] = mech
+                        
+                        ion = self._determine_ion(cm)
+                        if ion == 'non_specific':
+                            pass
+                            #mech['e'] = erev
+                        else:
+                            if not cellRule['secs'][section_name]['ions'].has_key(ion):
+                                cellRule['secs'][section_name]['ions'][ion] = {}
+                            ##cellRule['secs'][section_name]['ions'][ion]['e'] = erev
+                
+                for cm in cell.biophysical_properties.membrane_properties.channel_density_non_uniforms:
+                    
+                    for vp in cm.variable_parameters:
+                        if vp.parameter=="condDensity":
+                            iv = vp.inhomogeneous_value
+                            grp = vp.segment_groups
+                            path_vals = inhomogeneous_parameters[grp]
+                            expr = iv.value.replace('exp(','math.exp(')
+                            #print("variable_parameter: %s, %s, %s"%(grp,iv, expr))
+                            
+                            for section_name in seg_grps_vs_nrn_sections[grp]:
+                                path_start, path_end = inhomogeneous_parameters[grp][section_name]
+                                p = path_start
+                                gmax_start = pynml.convert_to_units('%s S_per_m2'%eval(expr),'S_per_cm2')
+                                p = path_end
+                                gmax_end = pynml.convert_to_units('%s S_per_m2'%eval(expr),'S_per_cm2')
+                                
+                                nseg = cellRule['secs'][section_name]['geom']['nseg'] if 'nseg' in cellRule['secs'][section_name]['geom'] else 1
+                                
+                                #print("   Cond dens %s: %s S_per_cm2 (%s um) -> %s S_per_cm2 (%s um); nseg = %s"%(section_name,gmax_start,path_start,gmax_end,path_end, nseg))
+                                
+                                gmax = []
+                                for fract in [(2*i+1.0)/(2*nseg) for i in range(nseg)]:
+                                    
+                                    p = path_start + fract*(path_end-path_start)
+                                    
+                                    
+                                    gmax_i = pynml.convert_to_units('%s S_per_m2'%eval(expr),'S_per_cm2')
+                                    #print("     Point %s at %s = %s"%(p,fract, gmax_i))
+                                    gmax.append(gmax_i)
+                                
+                                if cm.ion_channel=='pas':
+                                    mech = {'g':gmax}
+                                else:
+                                    mech = {'gmax':gmax}
+                                erev = pynml.convert_to_units(cm.erev,'mV')
+
+                                cellRule['secs'][section_name]['mechs'][cm.ion_channel] = mech
+
+                                ion = self._determine_ion(cm)
+                                if ion == 'non_specific':
+                                    mech['e'] = erev
+                                else:
+                                    if not cellRule['secs'][section_name]['ions'].has_key(ion):
+                                        cellRule['secs'][section_name]['ions'][ion] = {}
+                                    cellRule['secs'][section_name]['ions'][ion]['e'] = erev
+                                
                             
                 for cm in cell.biophysical_properties.membrane_properties.channel_density_ghks:
                     raise Exception("<channelDensityGHK> not yet supported!")
-                
-                for cm in cell.biophysical_properties.membrane_properties.channel_density_ghk2s:
-                    raise Exception("<channelDensityGHK2> not yet supported!")
-                
-                for cm in cell.biophysical_properties.membrane_properties.channel_density_non_uniforms:
-                    raise Exception("<channelDensityNonUniform> not yet supported!")
                 
                 for cm in cell.biophysical_properties.membrane_properties.channel_density_non_uniform_nernsts:
                     raise Exception("<channelDensityNonUniformNernst> not yet supported!")
@@ -904,17 +1023,26 @@ if neuromlExists:
                 
                 if self.verbose: print("Abstract cell: %s"%(isinstance(component_obj,BaseCell)))
                 
+                if hasattr(component_obj,'thresh'):
+                    threshold = pynml.convert_to_units(component_obj.thresh,'mV')
+                elif hasattr(component_obj,'v_thresh'):
+                    threshold = float(component_obj.v_thresh) # PyNN cells...
+                else:
+                    threshold = 0.0
+                    
+                
                 if not isinstance(component_obj,BaseCell):
                     popInfo['originalFormat'] = 'NeuroML2_SpikeSource' 
                     
                 cellRule = {'label': component, 
                             'conds': {'cellType': component, 
                                       'cellModel': component},  
-                            'sections': {}} # This parameter is required to distinguish NML2 "point processes" from artificial cells
+                            'secs': {}} # This parameter is required to distinguish NML2 "point processes" from artificial cells
 
                 soma = {'geom': {}, 'pointps':{}}  # soma properties
                 default_diam = 10
                 soma['geom'] = {'diam': default_diam, 'L': default_diam}
+                soma['threshold'] = threshold
                 
                 # TODO: add correct hierarchy to Schema for baseCellMembPotCap etc. and use this...
                 if hasattr(component_obj,'C'):
@@ -943,6 +1071,18 @@ if neuromlExists:
               
             self.gids[population_id] = [-1]*size
 
+        def _determine_ion(self, channel_density):
+            ion = channel_density.ion
+            if not ion:
+                if 'na' in channel_density.ion_channel.lower():
+                    ion = 'na'
+                elif 'k' in channel_density.ion_channel.lower():
+                    ion = 'k'
+                elif 'ca' in channel_density.ion_channel.lower():
+                    ion = 'ca'
+                else:
+                    ion = 'non_specific'
+            return ion
 
         def _convert_to_nrn_section_location(self, population_id, seg_id, fract_along):
             
@@ -1161,6 +1301,9 @@ if neuromlExists:
             preComp = nmlHandler.pop_ids_vs_components[prePop]
             
             from neuroml import Cell
+            '''
+            
+            No longer used in connections, defined in section on cell...
             
             if isinstance(preComp,Cell):
                 if len(preComp.biophysical_properties.membrane_properties.spike_threshes)>0:
@@ -1176,13 +1319,14 @@ if neuromlExists:
             elif hasattr(preComp,'v_thresh'):
                 threshold = float(preComp.v_thresh) # PyNN cells...
             else:
-                threshold = 0
+                threshold = 0.0'''
 
             for conn in nmlHandler.connections[projName]:
                 
                 pre_id, pre_seg, pre_fract, post_id, post_seg, post_fract, delay, weight = conn
                 
-                connParam = {'delay':delay,'weight':weight,'synsPerConn':1, 'sec':post_seg, 'loc':post_fract, 'threshold':threshold}
+                #connParam = {'delay':delay,'weight':weight,'synsPerConn':1, 'sec':post_seg, 'loc':post_fract, 'threshold':threshold}
+                connParam = {'delay':delay,'weight':weight,'synsPerConn':1, 'sec':post_seg, 'loc':post_fract}
                 
                 if ptype == 'electricalProjection':
 
@@ -1198,8 +1342,8 @@ if neuromlExists:
                                  'weight': weight,
                                  'synsPerConn': 1, 
                                  'sec': post_seg, 
-                                 'loc': post_fract, 
-                                 'threshold': threshold}
+                                 'loc': post_fract} 
+                                 #'threshold': threshold}
 
                 connParam['synMech'] = synapse